fixing a weirdness
