+++ src/ingest-pipeline/airflow/dags/export_and_backup/export_and_backup_plugin.py
@@ -1,6 +1,14 @@
-from export_and_backup.export_and_backup_plugin import ExportAndBackupPlugin
+from pathlib import Path
 
 
-class AlwaysPassesPlugin(ExportAndBackupPlugin):
+class ExportAndBackupPlugin:
     def run_plugin(self):
-        return "I passed!"
+        raise NotImplementedError()
+
+
+def export_and_backup_result_iter(plugin_dir, **kwargs) -> list:
+    plugin_dir = Path(plugin_dir)
+    plugins = list(plugin_dir.glob("*.py"))
+    if not plugins:
+        raise Exception(f"{plugin_dir}/*.py does not match any export_and_backup plugins")
+    return plugins
