+++ src/ingest-pipeline/airflow/dags/export_and_backup/plugins/always_passes_plugin.py
@@ -1,14 +1,6 @@
-from pathlib import Path
+from export_and_backup.export_and_backup_plugin import ExportAndBackupPlugin
 
 
-class ExportAndBackupPlugin:
+class AlwaysPassesPlugin(ExportAndBackupPlugin):
     def run_plugin(self):
-        raise NotImplementedError()
-
-
-def export_and_backup_result_iter(plugin_dir, **kwargs) -> list:
-    plugin_dir = Path(plugin_dir)
-    plugins = list(plugin_dir.glob("*.py"))
-    if not plugins:
-        raise Exception(f"{plugin_dir}/*.py does not match any export_and_backup plugins")
-    return plugins
+        return "I passed!"
